requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.15
django-cors-middleware==1.3.1
django-extensions==1.7.1
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.11.0
dnspython==1.16.0
drf-yasg==1.17.1
Flask==1.1.1
idna==2.9
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
packaging==20.1
protobuf==3.6.1
PyJWT==1.4.2
PyMySQL==0.9.3
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.10.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==0.16.0
